# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "École Platform API",
    "description": "EdTech SaaS platform API for K-12 schools in Morocco.\n\n## Domains\n- **IAM** — Authentication, authorization, invitations, recovery, 2FA\n- **ERP** — Classes, enrollments, teacher assignments, attendance\n- **LMS** — Courses, assignments, submissions, results, content, activities, assessments\n- **Billing** — Invoices and payments\n- **COM** — Notifications, consents, parent feed\n- **AI** — Writing assistance, recommendations, KPIs\n\n## Authentication\nAll protected endpoints require a `Bearer` access token in the `Authorization` header.\nTokens are obtained via `POST /auth/login`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/submissions": {
      "post": {
        "tags": [
          "lms-submissions"
        ],
        "summary": "Submit student work",
        "description": "Create a submission for an assignment.\n\nValidates:\n1. Assignment exists\n2. Course is in the same school\n3. INV-LMS-SUBMISSION: one active submission per student per assignment (idempotent)",
        "operationId": "create_submission_api_v1_submissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Submission record",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/submissions/{submission_id}/grade": {
      "post": {
        "tags": [
          "lms-submissions"
        ],
        "summary": "Grade a submission",
        "description": "Grade a student submission.\n\nValidates:\n1. Submission exists\n2. Assignment is in teacher's course (teacher owns the course)\n3. Submission must be in 'submitted' status\n4. Score <= assignment total_points\n5. Creates grade, updates submission status to 'graded'",
        "operationId": "grade_submission_api_v1_submissions__submission_id__grade_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Submission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Updated submission with grade",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results": {
      "get": {
        "tags": [
          "lms-results"
        ],
        "summary": "List student results",
        "description": "List graded results for a student.\n\nSTD: sees own results (student_id ignored, uses auth.user_id).\nPAR: must specify student_id, which must be a linked child.",
        "operationId": "list_results_api_v1_results_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of grades and results",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assessments/{assessment_id}/results": {
      "post": {
        "tags": [
          "lms-assessments"
        ],
        "summary": "Submit assessment result",
        "description": "Submit a result for an assessment (student self-report or auto-graded).\n\nValidates:\n1. Assessment exists and is published\n2. No duplicate result for same student+assessment (idempotent)",
        "operationId": "submit_assessment_result_api_v1_assessments__assessment_id__results_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assessment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentResultSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assessment result record",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentResultSubmitRequest": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assessment Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "assessment_id"
        ],
        "title": "AssessmentResultSubmitRequest"
      },
...
      "GradeRequest": {
        "properties": {
          "score": {
            "type": "number",
            "minimum": 0.0,
            "title": "Score"
          },
          "feedback_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Text"
          },
          "publish": {
            "type": "boolean",
            "title": "Publish",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "GradeRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubmissionCreateRequest": {
        "properties": {
          "assignment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignment Id"
          }
        },
        "type": "object",
        "required": [
          "assignment_id"
        ],
        "title": "SubmissionCreateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}